v. 0.6.13 - pip
Added getRN for getting fields and spots info
chenged getRNInfo (not compatible with the previous versions)
Added optimisation module with a few functions for pencil beam distribution calculation on a contour. This is still under development therefore not documented yet.
Updated documentation with getRN and getRNInfo functions.
A few bug fix.