v. 0.6.77 - pip
Removed pyamtrack from the dependencies. pyamtrack dose not work correctly on some MAC OS versions.

Fix: mapStructToImg - apparently, in some cases, the first contour of a structure was not mapped. Corrected this and tested against TPS volumes. The bug found thanks to Damian Borys

Fix: changed the default method of mapping structure to a voxelized 3D image from "centerInside" to "allInside". Changed this for mapStructToImg and getDVH. 

Updated documentation